+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2023-01-23
+
+*HTTP Parameter Pollution, bug fixes and more*
+
+### Added
+
+- HTTP Parameter Pollution module - Replaces array parameters with their last value (req.query must be set by a web framework)
+- README.md with module descriptions to [lib/modules](lib/modules/)
+- Issue templates and [CONTRIBUTING.md](CONTRIBUTING.md)
+
+### Changed
+
+- Fixed critical bug in "Block Tor Exit Nodes" module
+- Improved tests and updated examples
+- Code refactoring
+
 ## [0.3.1] - 2022-12-17
+
+*Security Update*
 
 ### Changed
 
@@ -10,7 +28,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.3.0] - 2022-11-06
 
-Hooks, log request method, improvements and bug fixes.
+*Hooks, log request method, improvements and bug fixes.*
 
 ### Added
 
@@ -28,7 +46,7 @@ Hooks, log request method, improvements and bug fixes.
 
 ## [0.2.0] - 2022-10-23
 
-The second beta release.
+*The second beta release.*
 
 ### Added
 
@@ -44,4 +62,4 @@ The second beta release.
 
 ## [0.1.0] - 2022-10-03
 
-This is the initial beta release.
+*This is the initial beta release.*